Summary of changes from v2.4.19-rc4 to v2.4.19-rc5 ============================================
<davem@redhat.com> (02/08/01 1.662)
	[PATCH] Correct openprom fix

	   <davem@redhat.com> (02/07/31 1.661)
	   	[PATCH] Add missing check to openprom driver

<akpm@zip.com.au> (02/08/01 1.663)
	[PATCH] disable READA

<marcelo@plucky.distro.conectiva> (02/08/01 1.664)
	Change EXTRAVERSION to -rc5

+++ drivers/sbus/char/openprom.c
@@ -334,7 +334,7 @@ static int copyin_string(char *user, size_t len, char **ptr)
 {
 	char *tmp;
 
-	if ((int)len < 0 || (int)(len + 1) < 0)
+	if ((ssize_t)len < 0 || (ssize_t)(len + 1) < 0)
 		return -EINVAL;
 
 	tmp = kmalloc(len + 1, GFP_KERNEL);
